Fix for empty buffer :Git command in Neovim
The issue appeared to be that Vim and NeoVim's `tabedit` command does not
accept and empty buffer for `%`. It appears that `%` was only used in
the command because it was convenient, though I am not positive about
that conclusion. In my testing seems to show that doing a `-tabnew`
instead of `-tabedit %` works just as well.